requirements.txt
matplotlib
numpy
pyserial